fix:validacion eliminar convenio

+++ app/Http/Controllers/Documentos/ConveniosController.php
@@ -57,46 +57,50 @@ class ConveniosController extends Controller
 
     public function updateConvenio(UpdateConvenioRequest $request, $id_convenio)
     {
-            try {
-                $request_form   = ['anios_arancel', 'valor_arancel', 'n_resolucion', 'fecha_resolucion', 'observacion', 'especialidad_id'];
+        try {
+            $request_form   = ['anios_arancel', 'valor_arancel', 'n_resolucion', 'fecha_resolucion', 'observacion', 'especialidad_id'];
 
-                $convenio = Convenio::find($id_convenio);
+            $convenio = Convenio::find($id_convenio);
 
-                if($convenio){
+            if ($convenio) {
 
-                    $update = $convenio->update($request->only($request_form));
+                $update = $convenio->update($request->only($request_form));
 
-                    $convenio->update([
-                        'usuario_update_id' => auth()->user()->id,
-                        'fecha_update'      => Carbon::now()->toDateTimeString()
-                    ]);
+                $convenio->update([
+                    'usuario_update_id' => auth()->user()->id,
+                    'fecha_update'      => Carbon::now()->toDateTimeString()
+                ]);
 
-                    $with = ['especialidad.perfeccionamiento.tipo', 'especialidad.centroFormador', 'especialidad.profesional'];
+                $with = ['especialidad.perfeccionamiento.tipo', 'especialidad.centroFormador', 'especialidad.profesional'];
 
-                    $convenio = $convenio->fresh($with);
+                $convenio = $convenio->fresh($with);
 
-                    if($update){
-                        return response()->json(array(true, $convenio));
-                    }else{
-                        return response()->json(false);
-                    }
-
+                if ($update) {
+                    return response()->json(array(true, $convenio));
+                } else {
+                    return response()->json(false);
                 }
-            } catch (\Exception $error) {
-                return response()->json($error->getMessage());
             }
+        } catch (\Exception $error) {
+            return response()->json($error->getMessage());
+        }
     }
 
     public function deleteConvenio($uuid)
     {
         try {
             $convenio = Convenio::where('uuid', $uuid)->first();
-            if($convenio){
-                $delete = $convenio->delete();
-                if($delete){
-                    return response()->json(true);
-                }else{
-                    return response()->json(false);
+            if ($convenio) {
+
+                if ($convenio->facturas != null && $convenio->facturas()->count() > 0) {
+                    return response()->json('passing_facturas');
+                } else {
+                    $delete = $convenio->delete();
+                    if ($delete) {
+                        return response()->json(true);
+                    } else {
+                        return response()->json(false);
+                    }
                 }
             }
         } catch (\Exception $error) {
